adding support for ZooDataset parameters

+++ fiftyone/zoo/__init__.py
@@ -71,7 +71,9 @@ def list_downloaded_zoo_datasets(base_dir=None):
     return downloaded_datasets
 
 
-def download_zoo_dataset(name, split=None, splits=None, dataset_dir=None):
+def download_zoo_dataset(
+    name, split=None, splits=None, dataset_dir=None, cleanup=True, **kwargs
+):
     """Downloads the dataset of the given name from the FiftyOne Dataset Zoo.
 
     Any dataset splits that already exist in the specified directory are not
@@ -93,6 +95,11 @@ def download_zoo_dataset(name, split=None, splits=None, dataset_dir=None):
         dataset_dir (None): the directory into which to download the dataset.
             By default, :func:`fiftyone.core.dataset.get_default_dataset_dir`
             is used
+        cleanup (True): whether to cleanup any temporary files generated during
+            download
+        **kwargs: optional keyword arguments to pass to the constructor of the
+            :class:`ZooDataset` for the dataset via the syntax
+            ``ZooDataset(**kwargs)`` before performing the download
 
     Returns:
         tuple of
@@ -100,9 +107,11 @@ def download_zoo_dataset(name, split=None, splits=None, dataset_dir=None):
         -   info: the :class:`fiftyone.zoo.ZooDatasetInfo` for the dataset
         -   dataset_dir: the directory containing the dataset
     """
-    zoo_dataset, dataset_dir = _parse_dataset_details(name, dataset_dir)
+    zoo_dataset, dataset_dir = _parse_dataset_details(
+        name, dataset_dir, **kwargs
+    )
     info = zoo_dataset.download_and_prepare(
-        dataset_dir, split=split, splits=splits
+        dataset_dir, split=split, splits=splits, cleanup=cleanup
     )
     return info, dataset_dir
 
@@ -271,7 +280,7 @@ def load_zoo_dataset_info(name, dataset_dir=None):
         raise OSError("Dataset '%s' not found at '%s'" % (name, dataset_dir))
 
 
-def get_zoo_dataset(name):
+def get_zoo_dataset(name, **kwargs):
     """Returns the :class:`ZooDataset` instance for the dataset with the given
     name.
 
@@ -280,6 +289,7 @@ def get_zoo_dataset(name):
 
     Args:
         name: the name of the zoo dataset
+        **kwargs: optional keyword arguments for ``ZooDataset(**kwargs)``
 
     Returns:
         the :class:`ZooDataset` instance
@@ -293,7 +303,20 @@ def get_zoo_dataset(name):
         datasets = all_datasets[source]
         if name in datasets:
             zoo_dataset_cls = datasets[name]
-            return zoo_dataset_cls()
+
+            try:
+                zoo_dataset = zoo_dataset_cls(**kwargs)
+            except Exception as e:
+                zoo_dataset_name = zoo_dataset_cls.__name__
+                raise ValueError(
+                    "Failed to construct zoo dataset instance using syntax "
+                    "%s(**kwargs); you may need to supply mandatory arguments "
+                    "to the constructor via `kwargs`. Please consult the "
+                    "documentation of `%s` to learn more"
+                    % (zoo_dataset_name, etau.get_class_name(zoo_dataset_cls),)
+                ) from e
+
+            return zoo_dataset
 
     raise ValueError("Dataset '%s' not found in the zoo" % name)
 
@@ -345,8 +368,8 @@ def _get_zoo_dataset_sources():
     return all_sources, has_default
 
 
-def _parse_dataset_details(name, dataset_dir):
-    zoo_dataset = get_zoo_dataset(name)
+def _parse_dataset_details(name, dataset_dir, **kwargs):
+    zoo_dataset = get_zoo_dataset(name, **kwargs)
 
     if dataset_dir is None:
         dataset_dir = fod.get_default_dataset_dir(zoo_dataset.name)
@@ -366,6 +389,7 @@ class ZooDatasetInfo(etas.Serializable):
         downloaded_splits (None): a dict of :class:`ZooDatasetSplitInfo`
             instances describing the downloaded splits of the dataset, if
             applicable
+        parameters (None): a dict of parameters for the dataset
         classes (None): a list of class label strings
     """
 
@@ -375,16 +399,20 @@ class ZooDatasetInfo(etas.Serializable):
         dataset_type,
         num_samples,
         downloaded_splits=None,
+        parameters=None,
         classes=None,
     ):
-        # Parse inputs
         if zoo_dataset.has_splits and downloaded_splits is None:
             downloaded_splits = {}
+
+        if parameters is None:
+            parameters = zoo_dataset.parameters
 
         self._zoo_dataset = zoo_dataset
         self._dataset_type = dataset_type
         self.num_samples = num_samples
         self.downloaded_splits = downloaded_splits
+        self.parameters = parameters
         self.classes = classes
 
     @property
@@ -453,6 +481,8 @@ class ZooDatasetInfo(etas.Serializable):
         _attrs = ["name", "zoo_dataset", "dataset_type", "num_samples"]
         if self.downloaded_splits is not None:
             _attrs.append("downloaded_splits")
+        if self.parameters is not None:
+            _attrs.append("parameters")
         if self.classes is not None:
             _attrs.append("classes")
 
@@ -502,6 +532,7 @@ class ZooDatasetInfo(etas.Serializable):
             dataset_type,
             d["num_samples"],
             downloaded_splits=downloaded_splits,
+            parameters=d.get("parameters", None),
             classes=d.get("classes", None),
         )
 
@@ -547,6 +578,13 @@ class ZooDataset(object):
     def name(self):
         """The name of the dataset."""
         raise NotImplementedError("subclasses must implement name")
+
+    @property
+    def parameters(self):
+        """An optional dict of parameters describing the configuration of the
+        zoo dataset when it was downloaded.
+        """
+        return None
 
     @property
     def supported_splits(self):
@@ -609,7 +647,9 @@ class ZooDataset(object):
         """
         return os.path.join(dataset_dir, "info.json")
 
-    def download_and_prepare(self, dataset_dir, split=None, splits=None):
+    def download_and_prepare(
+        self, dataset_dir, split=None, splits=None, cleanup=True
+    ):
         """Downloads the dataset and prepares it for use in the given
         directory.
 
@@ -624,6 +664,8 @@ class ZooDataset(object):
             splits (None): a list of splits to download, if applicable. If
                 neither ``split`` nor ``splits`` are provided, the full dataset
                 is  downloaded
+            cleanup (True): whether to cleanup any temporary files generated
+                during download
 
         Returns:
             the :class:`ZooDatasetInfo` for the dataset
@@ -712,12 +754,13 @@ class ZooDataset(object):
             info.write_json(info_path, pretty_print=True)
             logger.info("Dataset info written to '%s'", info_path)
 
-        # Cleanup scratch directory
-        etau.delete_dir(scratch_dir)
+        # Cleanup scratch directory, if necessary
+        if cleanup:
+            etau.delete_dir(scratch_dir)
 
         return info
 
-    def _download_and_prepare(self, dataset_dir, scratch_dir, splits):
+    def _download_and_prepare(self, dataset_dir, scratch_dir, split):
         """Internal implementation of downloading the dataset and preparing it
         for use in the given directory.
 
@@ -725,8 +768,8 @@ class ZooDataset(object):
             dataset_dir: the directory in which to construct the dataset
             scratch_dir: a scratch directory to use to download and prepare
                 any required intermediate files
-            splits: the list of splits to download, or None if the dataset does
-                not have splits
+            split: the split to download, or None if the dataset does not have
+                splits
 
         Returns:
             tuple of
